Tulsa: fix missing image
The original image was hot-linked from Wikipedia, and was deleted about
a month ago.

Found the same picture at another site:
http://greaterthan.securespsites.com/Images1/Art/Tulsa%20aerial%20392%20-%20Copy.jpg
It is however pretty big, so instead of linking, resized and optimized it as:

convert -size 640x640 input.jpg temp.jpg
jpegtran -optimize temp > Tulsa_aerial.jpg

Now it should be back as it was originally, with self-hosted image.