Rakefile: reorganize build related tasks
Make the default task just a build, to be able to run it more programically.
The previous default (monitoring mode) is now "preview" task.
Removed sass, as it seems there are no assets that need it (only have
CSS at the moment), and "jekyll --auto --server" can deal with monitoring
for changes