Bump Ruby and Gems
When I pulled the site down just now and followed the instructions to
get it running, nothing worked.

Given that the site was running super old versions of everything, I
figured it might be wise to bundle update to more recent versions. I
updated to Jekyll 3.8.x because 4.0.0 had some deprecations I didn't
want to figure out around Kramdown. The configuration options I removed
weren't needed anymore and the Rakefile was updated to reflect the new
way of invoking jekyll to serve the site (which fixed some deprecation
warnings.)

With this commit, the instructions in the README now work for me and
we're not running ancient versions of things.